requirements.txt
python-dotenv==1.0.0
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-spectacular==0.26.5
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
